Bugfix computation of cases with more than one target.
The code for computing the cases in which more than one target is indicated
had most likely a bug. Now there is an option to compute the network for
all the targets at once if the user wants it.
When computing one network for each target and then merging them in a final
single network, the resulting network may not be a proper tree, i.e. it
may contain feedback loops.